Remove blob helper methods from storage package
Removing various blob upload convenience methods that maintain some
internal logic that makes it hard to operate on these blobs afterwards.
I plan to refactor and move such methods to a `blobutil` package later
on. Fixes #71.
